backend/requirements.txt
# DiKaTech Backend Dependencies
# Compatible with Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML - Core (new package)
google-genai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
reportlab>=4.0.0
